// src/app/model/enums-model.ts
export interface Enums {
  certificazione: { name: string }[];
  laurea: { name: string }[];
  contratto: { name: string }[];
  diploma: { name: string }[];
  esito: { name: string }[];
  // certificazione: [
  //     {
  //         name: string
  //     },
  //     {
  //         name: string
  //     },
  //     {
  //         name: string
  //     }
  // ],
  // laurea: [
  //     {
  //         name: string
  //     },
  //     {
  //         name: string
  //     },
  //     {
  //         name: string
  //     },
  //     {
  //         name: string
  //     },
  //     {
  //         name: string
  //     },
  //     {
  //         name: string
  //     }
  // ],
  // contratto: [
  //     {
  //         name: string
  //     },
  //     {
  //         name: string
  //     },
  //     {
  //         name: string
  //     },
  //     {
  //         name: string
  //     }
  // ],
  // diploma: [
  //     {
  //         name: string
  //     },
  //     {
  //         name: string
  //     },
  //     {
  //         name: string
  //     },
  //     {
  //         name: string
  //     }
  // ],
  // esito: [
  //     {
  //         name: string
  //     },
  //     {
  //         name: string
  //     },
  //     {
  //         name: string
  //     },
  //     {
  //         name: string
  //     }
  // ]
}